Renamed fetch command to 'check'
