+++ src/commands/check.ts
